chore: removes container completely
